+++ KernelLand/Modules/Filesystems/Ext2/ext2_common.h
@@ -40,8 +40,8 @@ extern tVFS_Node	*Ext2_FindDir(tVFS_Node *Node, const char *FileName);
 extern int	Ext2_MkNod(tVFS_Node *Node, const char *Name, Uint Flags);
 extern int	Ext2_Link(tVFS_Node *Parent, tVFS_Node *Node, const char *Name);
 // --- Read ---
-extern Uint64	Ext2_Read(tVFS_Node *node, Uint64 offset, Uint64 length, void *buffer);
+extern size_t	Ext2_Read(tVFS_Node *node, off_t offset, size_t length, void *buffer);
 // --- Write ---
-extern Uint64	Ext2_Write(tVFS_Node *node, Uint64 offset, Uint64 length, const void *buffer);
+extern size_t	Ext2_Write(tVFS_Node *node, off_t offset, size_t length, const void *buffer);
 
 #endif
